Restructured the query for posts to index
Having duplicate queries is an invitation for errors. The limited query was missing $negative_restriction. The only reason I can see for having two separate queries was the "AND post.ID NOT IN (SELECT DISTINCT(doc) FROM $relevanssi_table)" part, which shouldn't be of any concern if the $relevanssi_table is empty anyway.

I also introduced the $limit_sql variable since I believe it is bad practice to reuse an integer variable as a string in the same code block.